refactor: separate talk editor logic, fix provider
