+++ .storybook/preview.tsx
@@ -1,8 +1,24 @@
+import { useState } from 'react';
+
 import type { Preview } from '@storybook/nextjs-vite';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 import '../src/app/globals.css';
 
+function StorybookQueryProvider({ children }: { children: React.ReactNode }) {
+  const [queryClient] = useState(() => new QueryClient());
+
+  return <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
+}
+
 const preview: Preview = {
+  decorators: [
+    (Story) => (
+      <StorybookQueryProvider>
+        <Story />
+      </StorybookQueryProvider>
+    ),
+  ],
   parameters: {
     nextjs: {
       appDirectory: true,
