+++ src/app/sosotalk/write/_components/sosotalk-post-editor/_components/post-editor-toolbar/index.ts
@@ -1,0 +1,2 @@
+export { PostEditorToolbar } from './post-editor-toolbar';
+export type { PostEditorToolbarAction, PostEditorToolbarProps } from './post-editor-toolbar.types';
