+++ src/app/sosotalk/write/_components/sosotalk-post-editor/_components/post-editor-toolbar/post-editor-toolbar.constants.ts
@@ -1,0 +1,25 @@
+import {
+  AlignCenter,
+  AlignLeft,
+  Bold,
+  ImagePlus,
+  Italic,
+  List,
+  ListOrdered,
+  Pilcrow,
+  Underline,
+} from 'lucide-react';
+
+import type { PostEditorToolbarAction } from './post-editor-toolbar.types';
+
+export const POST_EDITOR_TOOLBAR_ACTIONS: PostEditorToolbarAction[] = [
+  { label: '굵게', icon: Bold, command: 'bold', activeKey: 'bold' },
+  { label: '기울임', icon: Italic, command: 'italic', activeKey: 'italic' },
+  { label: '밑줄', icon: Underline, command: 'underline', activeKey: 'underline' },
+  { label: '본문', icon: Pilcrow, command: 'formatBlock', value: 'p' },
+  { label: '좌측 정렬', icon: AlignLeft, command: 'justifyLeft', activeKey: 'alignLeft' },
+  { label: '가운데 정렬', icon: AlignCenter, command: 'justifyCenter', activeKey: 'alignCenter' },
+  { label: '목록', icon: List, command: 'insertUnorderedList', activeKey: 'unorderedList' },
+  { label: '번호 목록', icon: ListOrdered, command: 'insertOrderedList', activeKey: 'orderedList' },
+  { label: '이미지 추가', icon: ImagePlus, isImage: true },
+];
