+++ src/app/sosotalk/write/_components/sosotalk-post-editor/hooks/use-sosotalk-post-editor.ts
@@ -7,6 +7,7 @@ import {
   SOSOTALK_POST_EDITOR_CONTENT_MAX_LENGTH,
   SOSOTALK_POST_EDITOR_TITLE_MAX_LENGTH,
 } from '../sosotalk-post-editor.constants';
+import type { ActiveFormats } from '../sosotalk-post-editor.internal-types';
 import type { SosoTalkPostEditorProps } from '../sosotalk-post-editor.types';
 import {
   calculateEffectiveContentLength,
@@ -15,16 +16,7 @@ import {
   toInitialHtml,
 } from '../sosotalk-post-editor.utils';
 
-type ActiveKey =
-  | 'bold'
-  | 'italic'
-  | 'underline'
-  | 'unorderedList'
-  | 'orderedList'
-  | 'alignLeft'
-  | 'alignCenter';
-
-type ActiveFormats = Record<ActiveKey, boolean>;
+import { useSosoTalkPostEditorImage } from './use-sosotalk-post-editor-image';
 
 const DEFAULT_ACTIVE_FORMATS: ActiveFormats = {
   bold: false,
@@ -59,10 +51,12 @@ export const useSosoTalkPostEditor = ({
   const [effectiveContentLength, setEffectiveContentLength] = useState(
     normalizePlainText(initialContent).length
   );
-  const [imageFile, setImageFile] = useState<File | null>(null);
-  // Preview URL is UI-only state for local rendering and should not be sent to the API.
-  const [imagePreviewUrl, setImagePreviewUrl] = useState(initialImageUrl);
   const [activeFormats, setActiveFormats] = useState<ActiveFormats>(DEFAULT_ACTIVE_FORMATS);
+  const { handleImageChange, handleImageRemove, imageFile, imagePreviewUrl } =
+    useSosoTalkPostEditorImage({
+      fileInputRef,
+      initialImageUrl,
+    });
 
   useEffect(() => {
     setTitle(initialTitle);
@@ -84,23 +78,6 @@ export const useSosoTalkPostEditor = ({
       editorRef.current.innerHTML = nextHtml;
     }
   }, [initialContent]);
-
-  useEffect(() => {
-    setImagePreviewUrl(initialImageUrl);
-  }, [initialImageUrl]);
-
-  useEffect(() => {
-    if (!imageFile) {
-      return;
-    }
-
-    const objectUrl = URL.createObjectURL(imageFile);
-    setImagePreviewUrl(objectUrl);
-
-    return () => {
-      URL.revokeObjectURL(objectUrl);
-    };
-  }, [imageFile]);
 
   const updateToolbarState = () => {
     if (!editorRef.current || typeof document === 'undefined') {
@@ -185,24 +162,6 @@ export const useSosoTalkPostEditor = ({
     if (command) {
       document.execCommand(command, false, value);
       syncEditorState();
-    }
-  };
-
-  const handleImageChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const nextFile = event.target.files?.[0] ?? null;
-    setImageFile(nextFile);
-
-    if (!nextFile) {
-      setImagePreviewUrl('');
-    }
-  };
-
-  const handleImageRemove = () => {
-    setImageFile(null);
-    setImagePreviewUrl('');
-
-    if (fileInputRef.current) {
-      fileInputRef.current.value = '';
     }
   };
