+++ src/app/sosotalk/write/_components/sosotalk-post-editor/sosotalk-post-editor.internal-types.ts
@@ -1,0 +1,10 @@
+export type ActiveFormatKey =
+  | 'bold'
+  | 'italic'
+  | 'underline'
+  | 'unorderedList'
+  | 'orderedList'
+  | 'alignLeft'
+  | 'alignCenter';
+
+export type ActiveFormats = Record<ActiveFormatKey, boolean>;
