+++ src/app/sosotalk/write/page.stories.tsx
@@ -1,10 +1,17 @@
-'use client';
-
 import type { Meta, StoryObj } from '@storybook/nextjs-vite';
 
 import { useAuthStore } from '@/store/auth-store';
 
 import SosoTalkWritePage from './page';
+
+const MOCK_LOGGED_IN_USER = {
+  id: 1,
+  email: 'user@example.com',
+  name: '김민주',
+  teamId: 'dallaem',
+  image:
+    'https://images.unsplash.com/photo-1500648767791-00dcc994a43e?q=80&w=300&auto=format&fit=crop',
+};
 
 const meta = {
   title: 'pages/sosotalk/write/page',
@@ -35,17 +42,7 @@ export const LoggedOut: Story = {
 export const LoggedIn: Story = {
   decorators: [
     (Story) => {
-      useAuthStore.setState({
-        user: {
-          id: 1,
-          email: 'user@example.com',
-          name: '김민주',
-          teamId: 'dallaem',
-          image:
-            'https://images.unsplash.com/photo-1500648767791-00dcc994a43e?q=80&w=300&auto=format&fit=crop',
-        },
-      });
-
+      useAuthStore.setState({ user: MOCK_LOGGED_IN_USER });
       return <Story />;
     },
   ],
